app/crawler/requirements.txt
# Requirements :
nltk==3.8.1
rake-nltk==1.0.6
beautifulsoup4==4.12.3
bs4==0.0.2
lxml==5.2.2
PyYAML==6.0.1
requests==2.32.3
regex==2024.5.15
mysql-connector-python==8.1.0